Fix com event hookup and ensure release
Declaring a variable on this level will cause the event not to fire when the variable is collected by the GC. You should declare this variable on the class level.

And I would consider getting this COM object and connecting to the event in the ADXAfterFormShow event and disconnecting from the event and releasing the COM object In the ADXAfterFormHide event.